backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
pytest==7.4.0
hypothesis==6.82.0
